[HERO/Clang] Print messages if HERCULES PREM passes are active
Console messages that indicate when HERCULES PREM passes are active and
when not are useful.  However, previously messages were always printed
if HERCULES PREM passes were *not* active, which confuses users that do
not want to use or do not know HERCULES PREM passes.  Additionally,
these messages break those upstream regression tests that match on
console output.

This commit changes the messages to be printed when HERCULES PREM passes
are active instead of when they are not active.