[OpenMP/HERO] Disable build of CUDA offloading plugin
The OpenMP offloading plugin for CUDA links to CUDA libraries.  When
those libraries are not available on the development machine, the LLVM
build fails with linker errors such as

    rtl.cpp:(.text._ZN12_GLOBAL__N_111DeviceRTLTyD2Ev+0x38f): undefined reference to `cuDevicePrimaryCtxRelease_v2'

As the CUDA offloading plugin is currently not used in HERO, we disable
it by default to avoid such problems.  Users can re-enable it by setting
`LIBOMPTARGET_CUDA_BUILD` to `ON` in their CMake configuration.